minor reorganization of IHashValue and HashValue classes
